// frontend/src/app/core/api/generated/model/patternTypeEnum.ts
/**
 * Chemistry Apps API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `toxicophore` - toxicophore * `privileged` - privileged
 */
export const PatternTypeEnum = {
    Toxicophore: 'toxicophore',
    Privileged: 'privileged'
} as const;
export type PatternTypeEnum = typeof PatternTypeEnum[keyof typeof PatternTypeEnum];