// frontend/src/app/core/api/generated/model/siteOverlapPolicyEnum.ts
/**
 * Chemistry Apps API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `last_block_wins` - last_block_wins
 */
export const SiteOverlapPolicyEnum = {
    LastBlockWins: 'last_block_wins'
} as const;
export type SiteOverlapPolicyEnum = typeof SiteOverlapPolicyEnum[keyof typeof SiteOverlapPolicyEnum];