// frontend/src/app/core/api/generated/model/smileitQuickProperties.ts
/**
 * Chemistry Apps API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Propiedades rápidas para decisiones de química medicinal.
 */
export interface SmileitQuickProperties { 
    molecular_weight: number;
    clogp: number;
    rotatable_bonds: number;
    hbond_donors: number;
    hbond_acceptors: number;
    tpsa: number;
    aromatic_rings: number;
}